Change update API  to task API (#1113)
* Change updates to tasks

* Complete implementation of asynchronous index CRUD

* Add client tests

* Add index tests

* Add documents test

* Improve js doc

* Add distinct attributes tests

* Add distrinct attributes tests

* Remove unused variable from tests

* Add dump tests

* Add filterable attributes tests

* Add ranking rules tests

* Add settings tests

* Add sortable attr tests

* Add stop words

* Fix types for Tasks

* Add search tests

* Finish tests compatibility

* Fix tasks issues

* Fix tests

* Fix task type

* Add await in return statement

* Fix tests

* Fix node12 jest tests

* Fix tests environments

* Remove obsolete test

* Fix tests

* Fix document tests

* Update update information in readme

* Add type check test

* Fix types of env

* Put back express tests to check for CORS issues

* Remove env from typescript type test

* Rename test file

* Fix typing

+++ tests/typed_search_tests.ts
@@ -1,13 +1,10 @@
-import { EnqueuedUpdate, ErrorStatusCode } from '../src/'
+import { EnqueuedTask, ErrorStatusCode } from '../src/'
 import {
   clearAllIndexes,
   config,
-  masterClient,
-  privateClient,
-  publicClient,
-  anonymousClient,
   BAD_HOST,
   MeiliSearch,
+  getClient,
 } from './meilisearch-test-utils'
 
 const index = {
@@ -73,30 +70,31 @@ afterAll(() => {
 })
 
 describe.each([
-  { client: masterClient, permission: 'Master' },
-  { client: privateClient, permission: 'Private' },
-  { client: publicClient, permission: 'Public' },
-])('Test on search', ({ client, permission }) => {
+  { permission: 'Master' },
+  { permission: 'Private' },
+  { permission: 'Public' },
+])('Test on search', ({ permission }) => {
   beforeAll(async () => {
+    const client = await getClient('Master')
     await clearAllIndexes(config)
-    await masterClient.createIndex(index.uid)
-    await masterClient.createIndex(emptyIndex.uid)
+
+    const task1 = await client.createIndex(index.uid)
+    await client.waitForTask(task1.uid)
+    const task2 = await client.createIndex(emptyIndex.uid)
+    await client.waitForTask(task2.uid)
+
     const newFilterableAttributes = ['genre', 'title']
-    const response: EnqueuedUpdate = await masterClient
+    const response: EnqueuedTask = await client
       .index<Movie>(index.uid)
       .updateFilterableAttributes(newFilterableAttributes)
-    expect(response).toHaveProperty('updateId', expect.any(Number))
-    await masterClient
-      .index<Movie>(index.uid)
-      .waitForPendingUpdate(response.updateId)
 
-    const { updateId } = await masterClient
-      .index<Movie>(index.uid)
-      .addDocuments(dataset)
-    await masterClient.index<Movie>(index.uid).waitForPendingUpdate(updateId)
+    await client.waitForTask(response.uid)
+    const { uid } = await client.index<Movie>(index.uid).addDocuments(dataset)
+    await client.waitForTask(uid)
   })
 
   test(`${permission} key: Basic search`, async () => {
+    const client = await getClient(permission)
     const response = await client.index<Movie>(index.uid).search('prince', {})
     expect(response.hits.length === 2).toBeTruthy()
     expect(response.offset === 0).toBeTruthy()
@@ -106,6 +104,7 @@ describe.each([
   })
 
   test(`${permission} key: Search with options`, async () => {
+    const client = await getClient(permission)
     const response = await client
       .index<Movie>(index.uid)
       .search('prince', { limit: 1 })
@@ -117,6 +116,7 @@ describe.each([
   })
 
   test(`${permission} key: Search with limit and offset`, async () => {
+    const client = await getClient(permission)
     const response = await client.index<Movie>(index.uid).search('prince', {
       limit: 1,
       offset: 1,
@@ -137,6 +137,7 @@ describe.each([
   })
 
   test(`${permission} key: Search with matches parameter and small croplength`, async () => {
+    const client = await getClient(permission)
     const response = await client.index<Movie>(index.uid).search('prince', {
       filter: 'title = "Le Petit Prince"',
       attributesToCrop: ['*'],
@@ -157,6 +158,7 @@ describe.each([
   })
 
   test(`${permission} key: Search with all options but not all fields`, async () => {
+    const client = await getClient(permission)
     const response = await client.index<Movie>(index.uid).search('prince', {
       limit: 5,
       offset: 0,
@@ -185,6 +187,7 @@ describe.each([
   })
 
   test(`${permission} key: Search with all options and all fields`, async () => {
+    const client = await getClient(permission)
     const response = await client.index<Movie>(index.uid).search('prince', {
       limit: 5,
       offset: 0,
@@ -211,6 +214,7 @@ describe.each([
   })
 
   test(`${permission} key: Search with all options but specific fields`, async () => {
+    const client = await getClient(permission)
     const response = await client.index<Movie>(index.uid).search('prince', {
       limit: 5,
       offset: 0,
@@ -247,6 +251,7 @@ describe.each([
   })
 
   test(`${permission} key: Search with specific fields in attributesToHighlight and check for types of number fields`, async () => {
+    const client = await getClient(permission)
     const response = await client.index<Movie>(index.uid).search('prince', {
       attributesToHighlight: ['title'],
     })
@@ -256,6 +261,7 @@ describe.each([
   })
 
   test(`${permission} key: Search with specific fields in attributesToHighlight and check for types of number fields`, async () => {
+    const client = await getClient(permission)
     const response = await client.index<Movie>(index.uid).search('prince', {
       attributesToHighlight: ['title', 'id'],
     })
@@ -265,6 +271,7 @@ describe.each([
   })
 
   test(`${permission} key: Search with filter and facetsDistribution`, async () => {
+    const client = await getClient(permission)
     const response = await client.index<Movie>(index.uid).search('a', {
       filter: ['genre=romance'],
       facetsDistribution: ['genre'],
@@ -275,6 +282,7 @@ describe.each([
   })
 
   test(`${permission} key: Search with filter with spaces`, async () => {
+    const client = await getClient(permission)
     const response = await client.index<Movie>(index.uid).search('h', {
       filter: ['genre="sci fi"'],
     })
@@ -283,6 +291,7 @@ describe.each([
   })
 
   test(`${permission} key: Search with multiple filter`, async () => {
+    const client = await getClient(permission)
     const response = await client.index<Movie>(index.uid).search('a', {
       filter: ['genre=romance', ['genre=romance', 'genre=romance']],
       facetsDistribution: ['genre'],
@@ -293,6 +302,7 @@ describe.each([
   })
 
   test(`${permission} key: Search with multiple filter and placeholder search using undefined`, async () => {
+    const client = await getClient(permission)
     const response = await client.index<Movie>(index.uid).search(undefined, {
       filter: ['genre = fantasy'],
       facetsDistribution: ['genre'],
@@ -303,6 +313,7 @@ describe.each([
   })
 
   test(`${permission} key: Search with multiple filter and placeholder search using NULL`, async () => {
+    const client = await getClient(permission)
     const response = await client.index<Movie>(index.uid).search(null, {
       filter: ['genre = fantasy'],
       facetsDistribution: ['genre'],
@@ -313,6 +324,7 @@ describe.each([
   })
 
   test(`${permission} key: Search on index with no documents and no primary key`, async () => {
+    const client = await getClient(permission)
     const response = await client.index(emptyIndex.uid).search('prince', {})
     expect(response.limit === 20).toBeTruthy()
     expect(response).toHaveProperty('processingTimeMs', expect.any(Number))
@@ -320,20 +332,25 @@ describe.each([
   })
 
   test(`${permission} key: Try to Search on deleted index and fail`, async () => {
-    await masterClient.index<Movie>(index.uid).delete()
+    const client = await getClient(permission)
+    const masterClient = await getClient('Master')
+    const { uid } = await masterClient.index<Movie>(index.uid).delete()
+    await masterClient.waitForTask(uid)
     await expect(
       client.index<Movie>(index.uid).search('prince')
     ).rejects.toHaveProperty('code', ErrorStatusCode.INDEX_NOT_FOUND)
   })
 })
 
-describe.each([{ client: anonymousClient, permission: 'Client' }])(
+describe.each([{ permission: 'No' }])(
   'Test failing test on search',
-  ({ client, permission }) => {
+  ({ permission }) => {
     beforeEach(async () => {
       await clearAllIndexes(config)
     })
+
     test(`${permission} key: Try Basic search and be denied`, async () => {
+      const client = await getClient(permission)
       await expect(
         client.index<Movie>(index.uid).search('prince')
       ).rejects.toHaveProperty(
